src/main/java/org/wecancodeit/reviews/storage/CategoryStorageJpaImpl.java
package org.wecancodeit.reviews.storage;

import org.springframework.stereotype.Service;
import org.wecancodeit.reviews.models.Category;
import org.wecancodeit.reviews.storage.repositories.CategoryRepository;

import java.util.Collection;

@Service
public class CategoryStorageJpaImpl implements CategoryStorage {

    CategoryRepository repository;

    public CategoryStorageJpaImpl(CategoryRepository repository) {
        this.repository = repository;
    }

    @Override
    public Collection<Category> getAll() {
        return (Collection<Category>) repository.findAll();
    }

    @Override
    public void store(Category category) {
        repository.save(category);
    }

    @Override
    public Category findCategoryByName(String name) {
        return repository.findByName(name).get();
    }
}
